refactor state management by using context api
