+++ app/home/(sections)/Toolbar.tsx
@@ -5,41 +5,25 @@ import IconButton from '@/app/_components/IconButton';
 import WhiteLoading from "/public/loadingWhite.svg";
 import { Snippet } from '../types';
 import { languages } from '../constants';
-import { handleUpdateSelectedSnippetLanguage, handleUpdateSnippetLanguage } from '../_utility/updateData';
+import { handleUpdateCollectionTitle, handleUpdateSelectedSnippetLanguage, handleUpdateSnippetLanguage } from '../_utility/updateData';
 import { DropdownMenu, DropdownMenuContent, DropdownMenuTrigger } from '@radix-ui/react-dropdown-menu';
 import { ChevronDown, CirclePlus, Earth, Folder, Share2, Trash2 } from 'lucide-react';
 import MyDropdownMenu from '@/app/_components/DropdownMenu';
+import { useAppContext } from '@/app/_context/AppContext';
+import { handleAddSnippet } from '../_utility/addData';
+import { handleMoveSnippets, handleShare } from '../_utility/otherHandlers';
+import { handleDeleteSnippets } from '../_utility/deleteData';
 
 interface ToolbarProps {
-  activeCollection: any;
-  loadingAddSnippet: boolean;
-  handleUpdateCollectionTitle: (event: React.ChangeEvent<HTMLInputElement>, collection: any, setCollections: any, setActiveCollection: any) => void;
-  handleAddSnippet: () => void;
-  handleShare: () => void;
-  handleDeleteSnippets: (ids: any[], setSnippets: any) => void;
-  selectedSnippetsId: any[];
-  handleMoveSnippets: () => void;
-  setCollections: (collections: any[]) => void;
-  setActiveCollection: (collection: any) => void;
-  setSnippets: Dispatch<SetStateAction<Snippet[]>>;
+  scrollToBottom: () => void;
 }
 
-const Toolbar: React.FC<ToolbarProps> = ({
-  activeCollection,
-  loadingAddSnippet,
-  handleUpdateCollectionTitle,
-  handleAddSnippet,
-  handleShare,
-  handleDeleteSnippets,
-  selectedSnippetsId,
-  handleMoveSnippets,
-  setCollections,
-  setActiveCollection,
-  setSnippets
-}) => {
+const Toolbar: React.FC<ToolbarProps> = ({scrollToBottom}) => {
+
+  const {activeCollection, loadingAddSnippet, selectedSnippetsId, setSnippets} = useAppContext();
 
   const [selectedLanguage, setSelectedLanguage] = useState('python');
-  
+
   return (
   <div className="p-2 border-b border-black bg-neutral-900 text-white">
     <p className="text-2xl font-bold flex">
@@ -49,7 +33,7 @@ const Toolbar: React.FC<ToolbarProps> = ({
         type="text"
         disabled={!activeCollection}
         value={activeCollection?.title}
-        onChange={(event) => handleUpdateCollectionTitle(event, activeCollection, setCollections, setActiveCollection)}
+        onChange={(event) => handleUpdateCollectionTitle(event)}
       />
       {loadingAddSnippet && <span className="pl-2">
         <Image
@@ -61,18 +45,20 @@ const Toolbar: React.FC<ToolbarProps> = ({
       </span>}
     </p>
     <div className="flex space-x-4">
-      <IconButton icon={<CirclePlus/>} text="New code snippet" onClick={handleAddSnippet} disabled={!activeCollection} isDark />
+      <IconButton icon={<CirclePlus/>} text="New code snippet" onClick={() => {
+        handleAddSnippet(scrollToBottom);
+      }} disabled={!activeCollection} isDark />
       {activeCollection?.shared
         ? <IconButton iconColor="green" textColor="green" icon={<Earth />} text="Public" isDark onClick={handleShare} />
         : <IconButton icon={<Share2 />} text="Share" isDark disabled={!activeCollection} onClick={handleShare} />}
-      <IconButton icon={<Trash2 />} text="Delete" isDark disabled={selectedSnippetsId.length === 0} onClick={() => handleDeleteSnippets(selectedSnippetsId, setSnippets)} />
+      <IconButton icon={<Trash2 />} text="Delete" isDark disabled={selectedSnippetsId.length === 0} onClick={() => handleDeleteSnippets()} />
       <IconButton icon={<Folder/>} text="Move" isDark disabled={selectedSnippetsId.length === 0} onClick={handleMoveSnippets} />
       <MyDropdownMenu buttonText={selectedLanguage} disabled={selectedSnippetsId.length === 0}>
         <div className="text-black h-44 grid grid-cols-1 w-24 bg-slate-100 overflow-y-auto">
           {languages.map((lang, index) => (
             <button key={index} onClick={() => {
               setSelectedLanguage(lang);
-              handleUpdateSelectedSnippetLanguage(lang, setSnippets, selectedSnippetsId)
+              handleUpdateSelectedSnippetLanguage(lang)
             }}>
               <p className="hover:bg-slate-300 p-1">{lang}</p>
             </button>
