+++ app/home/_utility/deleteData.ts
@@ -1,14 +1,20 @@
 import supabase from "@/app/utils/supabase";
 import { Collection, Snippet } from "../types";
 import { Dispatch, SetStateAction } from "react";
+import { useAppContext } from "@/app/_context/AppContext";
 
-export const handleDeleteCollection = async (value: Collection, setShowOverlayMenuPage: (value: boolean) => void, setCurrentOverlayMenuPage: (page: string) => void, setToDeleteCollection: (collection: Collection) => void) => {
+export const handleDeleteCollection = async (
+    value: Collection,
+    setCurrentOverlayMenuPage: (currentOverlayMenuPage: string) => void,
+    setShowOverlayMenuPage: (showOverlayMenuPage: boolean) => void,
+) => {
     setShowOverlayMenuPage(true);
     setCurrentOverlayMenuPage("deleteCollectionConfirmation");
-    setToDeleteCollection(value);
+    // setToDeleteCollection(value);
 };
 
-export const handleDeleteSnippets = async (selectedSnippetsId: number[], setSnippets: Dispatch<SetStateAction<Snippet[]>>) => {
+export const handleDeleteSnippets = async () => {
+    const {selectedSnippetsId, setSnippets, snippets} = useAppContext();
     selectedSnippetsId.map(async (itemId) => {
         const { error } = await supabase
             .from('snippet')
@@ -18,11 +24,10 @@ export const handleDeleteSnippets = async (selectedSnippetsId: number[], setSnip
         if (error) {
             console.log(error);
         } else {
-            setSnippets((prevItems) => (
-                prevItems.filter((item) => (
+            setSnippets(snippets.filter((item) => (
                     !selectedSnippetsId.includes(item.id)
                 ))
-            ));
+            );
         }
     });
 };
