+++ app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
 import "./globalicons.css";
+import { AppProvider } from "./_context/AppProvider";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -17,7 +18,9 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
+      <AppProvider>
       <body className={inter.className}>{children}</body>
+      </AppProvider>
     </html>
   );
 }
